Haciendo uso de excepciones para manejo de errores

+++ server/Interpreter/Expresiones/Operaciones/Aritmeticas.py
@@ -50,7 +50,7 @@ class Aritmeticas(Operacion):
 
         elif self.operador == Operador.DIV:
             if tipo_exp1 == Tipos.INT64 and tipo_exp2 == Tipos.INT64:
-                    return int(self.exp1.getValor(driver, ts) / self.exp2.getValor(driver, ts))
+                    return self.exp1.getValor(driver, ts) / self.exp2.getValor(driver, ts)
             elif tipo_exp1 == Tipos.FLOAT64 and tipo_exp2 == Tipos.FLOAT64:
                     return self.exp1.getValor(driver, ts) / self.exp2.getValor(driver, ts)
             else:
